requirements.txt
datasets
qdrant-client
sentence-transformers
torch
transformers
gradio
pyyaml
tqdm